♻️ refactor: restructure services and update import paths
