+++ apps/blog/services/__init__.py
@@ -1,0 +1,1 @@
+from content_generator import ContentGenerator
